requirements.txt
# Core dependencies
flask>=2.3.0
gunicorn>=21.2.0
werkzeug>=3.0.0
python-dotenv>=0.19.1

# Google API dependencies
google-api-python-client>=2.31.0
google-auth>=2.3.0
google-auth-oauthlib>=0.4.6
google-analytics-data>=0.17.1
google-ads>=22.1.0

# Data processing
numpy>=1.26.0
pandas>=2.1.0

# Development dependencies
pytest>=7.1.1
pytest-cov>=3.0.0
flake8>=4.0.1
black>=22.3.0